+++ src/main/java/net/geforcemods/securitycraft/network/server/SetSentryMode.java
@@ -1,6 +1,8 @@
 package net.geforcemods.securitycraft.network.server;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import io.netty.buffer.ByteBuf;
 import net.geforcemods.securitycraft.entity.sentry.Sentry;
@@ -14,26 +16,30 @@ import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler;
 import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 
 public class SetSentryMode implements IMessage {
-	private BlockPos pos;
-	private int mode;
+	private List<Info> sentriesToUpdate;
 
 	public SetSentryMode() {}
 
-	public SetSentryMode(BlockPos sentryPos, int mode) {
-		pos = sentryPos;
-		this.mode = mode;
+	public SetSentryMode(List<Info> sentriesToUpdate) {
+		sentriesToUpdate.removeIf(Objects::isNull);
+		this.sentriesToUpdate = sentriesToUpdate;
 	}
 
 	@Override
 	public void fromBytes(ByteBuf buf) {
-		pos = BlockPos.fromLong(buf.readLong());
-		mode = buf.readInt();
+		int size = buf.readInt();
+
+		sentriesToUpdate = new ArrayList<>();
+
+		for (int i = 0; i < size; i++) {
+			sentriesToUpdate.add(Info.read(buf));
+		}
 	}
 
 	@Override
 	public void toBytes(ByteBuf buf) {
-		buf.writeLong(pos.toLong());
-		buf.writeInt(mode);
+		buf.writeInt(sentriesToUpdate.size());
+		sentriesToUpdate.forEach(info -> info.write(buf));
 	}
 
 	public static class Handler implements IMessageHandler<SetSentryMode, IMessage> {
@@ -43,15 +49,36 @@ public class SetSentryMode implements IMessage {
 				EntityPlayer player = context.getServerHandler().player;
 				World level = player.world;
 
-				if (level.isBlockLoaded(message.pos)) {
-					List<Sentry> sentries = level.<Sentry>getEntitiesWithinAABB(Sentry.class, new AxisAlignedBB(message.pos));
+				for (Info info : message.sentriesToUpdate) {
+					if (level.isBlockLoaded(info.pos)) {
+						List<Sentry> sentries = level.<Sentry>getEntitiesWithinAABB(Sentry.class, new AxisAlignedBB(info.pos));
 
-					if (!sentries.isEmpty() && sentries.get(0).isOwnedBy(player))
-						sentries.get(0).toggleMode(player, message.mode, false);
+						if (!sentries.isEmpty() && sentries.get(0).isOwnedBy(player))
+							sentries.get(0).toggleMode(player, info.mode, false);
+					}
 				}
 			});
 
 			return null;
+		}
+	}
+
+	public static class Info {
+		private final BlockPos pos;
+		private final int mode;
+
+		public Info(BlockPos pos, int mode) {
+			this.pos = pos;
+			this.mode = mode;
+		}
+
+		public static Info read(ByteBuf buf) {
+			return new Info(BlockPos.fromLong(buf.readLong()), buf.readInt());
+		}
+
+		public void write(ByteBuf buf) {
+			buf.writeLong(pos.toLong());
+			buf.writeInt(mode);
 		}
 	}
 }