+++ src/main/java/net/geforcemods/securitycraft/network/server/ToggleBlockPocketManager.java
@@ -1,11 +1,18 @@
 package net.geforcemods.securitycraft.network.server;
 
 import io.netty.buffer.ByteBuf;
+import net.geforcemods.securitycraft.SCContent;
+import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.blockentities.BlockPocketManagerBlockEntity;
+import net.geforcemods.securitycraft.network.client.BlockPocketManagerFailedActivation;
 import net.geforcemods.securitycraft.util.LevelUtils;
-import net.minecraft.entity.player.EntityPlayer;
+import net.geforcemods.securitycraft.util.PlayerUtils;
+import net.geforcemods.securitycraft.util.Utils;
+import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.text.TextComponentTranslation;
+import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler;
@@ -18,11 +25,11 @@ public class ToggleBlockPocketManager implements IMessage {
 
 	public ToggleBlockPocketManager() {}
 
-	public ToggleBlockPocketManager(BlockPocketManagerBlockEntity te, boolean enabling, int size) {
+	public ToggleBlockPocketManager(BlockPocketManagerBlockEntity te, boolean enabling) {
 		pos = te.getPos();
 		dimension = te.getWorld().provider.getDimension();
 		this.enabling = enabling;
-		this.size = size;
+		size = te.getSize();
 	}
 
 	@Override
@@ -45,19 +52,29 @@ public class ToggleBlockPocketManager implements IMessage {
 		@Override
 		public IMessage onMessage(ToggleBlockPocketManager message, MessageContext ctx) {
 			LevelUtils.addScheduledTask(ctx.getServerHandler().player.world, () -> {
-				EntityPlayer player = ctx.getServerHandler().player;
+				EntityPlayerMP player = ctx.getServerHandler().player;
 				World world = player.world;
 				TileEntity te = world.getTileEntity(message.pos);
 
 				if (te instanceof BlockPocketManagerBlockEntity && ((BlockPocketManagerBlockEntity) te).isOwnedBy(player)) {
+					TextComponentTranslation feedback;
+
 					((BlockPocketManagerBlockEntity) te).setSize(message.size);
 
 					if (message.enabling)
-						((BlockPocketManagerBlockEntity) te).enableMultiblock();
+						feedback = ((BlockPocketManagerBlockEntity) te).enableMultiblock();
 					else
-						((BlockPocketManagerBlockEntity) te).disableMultiblock();
+						feedback = ((BlockPocketManagerBlockEntity) te).disableMultiblock();
+
+					if (feedback != null) {
+						if (message.enabling && !((BlockPocketManagerBlockEntity) te).isEnabled())
+							SecurityCraft.network.sendTo(new BlockPocketManagerFailedActivation(message.pos), player);
+
+						PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.blockPocketManager), feedback, TextFormatting.DARK_AQUA, false);
+					}
 				}
 			});
+
 			return null;
 		}
 	}